--- Pattern3.java
@@ -1,15 +1,0 @@
-class Pattern3 
-{
-	public static void main(String[] args) 
-	{
-		for(int row = 1; row <= 3; row++)
-		{
-			for(int col = 5; col >=1; col--)
-			{
-
-			System.out.print(col+" ");
-			}
-			System.out.println();
-		}
-	}
-}
